requirements.txt
streamlit
Pillow
google-api-python-client
google-auth
google-auth-oauthlib
gspread
pandas
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1